added missing @override

+++ org.faktorips.fl/src/org/faktorips/fl/parser/FlParserTokenManager.java
@@ -864,7 +864,7 @@ public class FlParserTokenManager implements FlParserConstants {
                     }
                 } while (i != startsAt);
             } else {
-                int hiByte = (int)(curChar >> 8);
+                int hiByte = (curChar >> 8);
                 int i1 = hiByte >> 6;
                 long l1 = 1L << (hiByte & 077);
                 int i2 = (curChar & 0xff) >> 6;
